% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\docType{package}
\name{spantest-package}
\alias{spantest-package}
\title{Package imports}
\description{
Import functions from other packages
}
\author{
\strong{Maintainer}: Benjamin Seguin \email{benjamin.seguin@outlook.fr}

Authors:
\itemize{
  \item David Ardia \email{david.ardia.ch@gmail.com}
}

Other contributors:
\itemize{
  \item Rosnel Sessinou \email{srosnels@gmail.com} [contributor]
}

}
